remove specialized base64 feature
catzcod's modification allowed the decoding of content that was *served* as base64 text, as opposed to being encoded by the har creator. it's convenient, but it's a bit too specific for a main fork release, so i've removed it. may add it back some day if it's a very common scenario, though.